*: Add logic to teardown bootstrap
Add bootstrapteardown/teardown.go which watches on clusterversions.
The installer watches on clusterversions to wait for the installation
to complete. We use the same mechanism so removing the bootstrap from
ceo and removing the resources in installer are both triggered on
clusterversions reporting install complete and are executed at about
the same time preventing excessive etcd logging.