*: Add hostetcdendpointcontroller
The controller watches on cluster etcd CRD, endpoints and pods in
etcd namespace. It reconciles by running a diff on actual number of
members observed by etcd api with members in the endpoint list.

Querying the etcd membership is necessary here because all the
controllers in this project depends on the endpoints being the source of
truth. If the endpoint list is based on other things like
cluster.members data in etcd, it creates a cyclic dependency and is
not necessarily accurate.

hostendpointcontroller: fix typo