*: remove service-ca.crt and attempt to sync all required resources
The rest.InClusterConfig() only uses token and ca.crt in discovery
init container.

It was observed that service-ca.crt was the last file to be synced.
Since the loop would error out, it would take a long time for token to
be synced, which is the true dependency of discovery init container.

This PR removes service-ca.crt and fixes the logic by attempting to
sync all the resources even if previous one errored out.

This is expected to significantly reduce the bootstrap time.