Stop reconciling in response to global configmap events
Before this change, the ClusterMemberController performed a full sync in response
to any watch event associated with a configmap in any namespace in the cluster.
This causes (at a minimum) unnecessary etcd client traffic.

This patch changes the controller to stop reacting to configmaps entirely and instead
react to changes to nodes, which was probably the original intent.