Revert "remove dead host-etcd-2 service"
This reverts commit 7f142f69b1fe95ffb033526256d9498a6c8dadc5.

https://github.com/kubernetes/kubernetes/issues/6877
https://github.com/kubernetes/kubernetes/pull/7821
https://github.com/kubernetes/kubernetes/pull/7821#issuecomment-99344889

Deleting the service subjects the host-etcd-2 endpoint to deletion by
the Kube endpoints controller at random times now or in the future.

The service must remain to protect the existence of the endpoint given
the current design of Kube.

By deleting the service, we risk bootstrap failure through information
loss when the endpoint is deleted during bootstrapping, and churning
unnecessary endpoint revisions by recreating an endpoint Kube wants
to delete.