Replace etcd endpoint representation with configmap
Kube service design asserts `endpoint` resources cannot exist without a
corresponding `service` resource, and Kube will actively delete the endpoint
when the service is deleted or if Kube detects the endpoint is a "stray".

The operator needs to:

1. Manage etcd endpoint state atomically.
2. Maintain exclusive ownership of the etcd endpoint state resource.

Altogether this makes the `endpoint` resource inappropriate for the task. The
competition between the operator and the Kube endpoints controller to manage the
endpoint has led to instability.

To resolve the problems, persist etcd endpoint state in a `configmap`.

Maintain compatibility by continuing to write the `endpoint`, and update
consuming components to prefer the `configmap` over the `endpoint`.

Also requires:
https://github.com/openshift/cluster-kube-apiserver-operator/pull/859
https://github.com/openshift/cluster-openshift-apiserver-operator/pull/364