Defrag etcd on startup
This change introduces an etcd defrag on startup during pod init.

Performing the defrag on startup is a low risk way to establish a minimal
guarantee that defrag will occur automatically under at least one condition
(during an upgrade). This changes is not intended to replace the need for live
defrags during steady state operations.