Reduce superfluous leader change logging
Before this patch, leader change stats were logged at V(0) even when
the value is zero, which isn't useful. We already record an event when
the value is greater than zero.

This patch changes the leader change log to V(4) to reduce useless
log messages in the steady state while preserving the ability to enable
the log through operator configuration.