Allow rollbackcopier pod to be disabled
Allow the `rollbackcopier` container in the etcd pod to to be disabled using a
`openshift.io/disable-rollbackcopier` annotation. The `rollbackcopier` container
is only present in versions 4.4 and 4.5. This patch is implemented in such a way
that the disable function can be introduced in the latest release and
backported. The new code should effectively no-op in releases where
`rollbackcopier` is not present.