Use library-go for guard controller
We are the in the process of having guard
using PDB for all control plane components.
Making the current etcd quorum guard generic
enough and moving it to library-go is the
first step which is handled via
openshift/library-go#1095

This commit removes the existing quorum guard
controller from the etcd operator and use the
controller from library-go. The functionality
should remain as is because we have hooks
for the deployment to maintain the number
of replicas. The PDB & deployment won't be
created in case of SNO.