changes the signature of the MemberRemove method to accept a memberID (uint64) not a Name (string)
previously the method required contacting with the etcd cluster to get the current list of members just to find an ID of the member that matches the given name.
however, all current callers (i.e. the bootstrapteardown, member removal controllers) of this method already have an ID of the member they want to remove.
in addition to that, it is easier to deal with an ID instead of a Name because the latter might be sometimes empty (i.e. member hasn't been started) which leads to awkward error handling in the operator.