Replace quorumguard and add readyz server
This commit replaces the etcd-operator's quorumguard controller with the
staticpod controller's guard controller to ensure guard pods are created as
static pods and avoids the possibility of two guard pods being deployed on the same node.
Part of this change also adds an http-get server that serves /readyz and /healthz endpoints
on port 9980 to serve readiness probes for the etcd and guard pods.