no pending on etcdHighNumberOfLeaderChanges 1st 1h
during initial cluster roll-out there is enough churn that etcd
leadership will change and these alerts would result. This can
result of normal cluster installation and the alerts could cause
false concern of some problem.

this is happening more frquently in CI with OVN as the CNI,
especially in the GCP cloud (for whatever reason) and results
in extra failures in CI, as well as alarms customers would
see.

This should wait for an hour after "cluster_version" goes
initial and should provide enough time before the alert would
be much more meaningful if it came.

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>