etcd_perf file left after fio test when restoring
This commit deletes "etcd_perf" files that could present if someone has ran a fio test under /var/lib/etcd
like specified in the OpenShift docs about the recommended host practices.

As this file is known as non essential, it could be freely deleted by anticipation by the restore pod when
restoring a cluster to a previous state.

The problem here appears when restoring an Openshift cluster in a previous state following the documentation.

Everything went fine until the restauration and restart of the kubelet (Step 8.) because the pod that is in
charge of executing the db snapshot restore step isn't able to restore due to that file left behind by fio
perf test. The commit context do not show the 4th line after the modification where the script "exit 1".

This fix that problem, It isn't may be not acceptable to handle every files that could land here, but as
this fio's file comes from the documentation of recommended host practices, it could be worth it to handle
at least file issued from the documentation procedure.

Change-Id: Ib6d9e522bd68fff2aba437fe0ae3ec1bd7260baf
Signed-off-by: Romain Forlot <rforlot@yahoo.com>