Prefer ipv6 on bootstrap in v6-primary dual stack
Previously we only supported v4-primary dual stack, so it was correct
to always use ipv4 unless we were single stack v6. Now we need to add
support for v6-primary dual stack so we need to be more nuanced about
ip selection.

This change looks at the first CIDR to determine whether the cluster
should be v4 or v6 primary. The other logic stays essentially the
same since it was already set up to look for ipv6 addresses when
required.