OCPBUGS-18984: remove UpgradeBackupController
We determined that the UpgradeBackupController is not maintainable
anymore due to missing retention (frequently filling up CP node disks)
and incorrect behavior during z upgrades.

This PR removes the whole controller, as a joint decision from the team, and depending
change on CVO side in [CVO#968](openshift/cluster-version-operator#968).

This controller was never advertised as functionality to customers, it
was merely a last-resort backup solution for the etcd storage migration
between OCP 4.9/4.10 and should've been removed in 4.11.