ETCD-607: gate leaf cert generation
This implements a gate to avoid triggering leaf cert generation in the
same static pod revision as an update to the signer certificates (and
their respective bundles).

This PR also removes the openshift-config dependency on signer
certificates. With 4.17 we should be safe to remove any reference, as
the newly generated signer certificate in openshift-etcd should already
be distributed and ready for consumption.

In plain words: an upgrade to 4.17 will automatically rotate your etcd
signer certificates for the first time and every refresh period happily
ever after.

Signed-off-by: Thomas Jungblut <tjungblu@redhat.com>