ETCD-656: add cluster identifier to revision.json
This is required because the cluster id is not stored in the raft
hardstate. It proved difficult to re-generate this from within the
initial cluster discovery binary based on environment variables,
so we're storing it more explicitly in the revision.json.

This also fixes a flake when the leader is terminated in the unit tests.