OCPEDGE-1498: support updates to external etcd
The EnableEtcdContainer unsupported config override removes the etcd
container from the static pod definition when set to FALSE.

This change extends that behavior by writing a separate Pod definition
containing **only** the etcd container, using the same source from which
it was removed, but placing it in a location that does not trigger pod
creation.

This allows the external etcd manager to run etcd independently while
ensuring it has a reliable source for updates from CEO (e.g., new image,
path to certificates, etc.).