fix: remove shared informer for arbiter deployments
The previous implementation of a shared informer using multiple selectors caused a race condition that made master node additions not show up when deploying arbiter clusters. It's best to watch all the nodes, this should only be done for arbiter deployments, and filter at the lister level to keep things working as expected as well as giving us the OR filtering we were looking to achieve with the previous implementation.

Signed-off-by: ehila <ehila@redhat.com>