OCPBUGS-44570: add missing informer syncs
During apiserver/etcd downtime we may have problematic error messages
like "openshift-etcd not found". This is caused by the ns lister not
being synchronized yet, thus having an internal empty list to serve
from. Queries will return not found for namespaces that definitely
exists for this brief period of time.

This fix cleans up several informers that we have missed and cleans up
some that we don't need.

Be extra careful when reviewing, this PR was created with AI.

Signed-off-by: Thomas Jungblut <tjungblu@redhat.com>