TNF: Use operator condition instead of job status
The problem is that the job status isn't persistent. On CEO updates,
the job is recreated. That means that we don't know if pacemaker is
running the etcd container already or not, until the job condition is
set. On initial cluster installation, pacemaker's etcd isn't running
yet, but on updates it is. In the latter case we have to prevent that
CEO re-adds its own etcd container.
Using an operator condition is persistent and doesn't need workarounds
for handling both initial installation and CEO updates.

Signed-off-by: Marc Sluiter <msluiter@redhat.com>