Moved pacemaker status gathering to a cronjob with elevated permissions.
The goal of this change was to ensure that the CEO permissions for the pcs status cronjob
were elevated enough to be able to gather pcs status output. One future improvement would
be to move this out of the CEO entirely and have the status gathering be done by an
alert agent that posts regular updates to the CEO.

To alleviate the complexity of a future transition to this model, I've structured the code
so that the healthcheck controller doesn't care where the posts to the PacemakerStatus CRD
come from - it just monitors it for updates. In the future, we can write an alert agent
that just posts new PacemakerStatus updates as they happen, and the CEO can retire the
cronjob entire.